requirements.txt
packaging==23.2
langgraph==.0.0.32
langchain_openai==0.1.1
langchain==0.1.13
json5
python-dotenv
